Change time to live in Foodstuff to use longs

+++ src/gartham/c10ver/economy/items/utility/foodstuffs/Sandwich.java
@@ -9,7 +9,7 @@ public class Sandwich extends Foodstuff {
 
 	public static final String ITEM_TYPE = "sandwich", ITEM_NAME = "Sandwich", ITEM_ICON = ":sandwich:";
 	public static final BigDecimal ITEM_MULT = BigDecimal.valueOf(25, 2);// 0.25 (25, shifted to the right twice.)
-	public static final BigInteger ITEM_TTL = sec(150);// 2.5min
+	public static final long ITEM_TTL = sec(150);// 2.5min
 
 	{
 		setItemName(ITEM_NAME);
